Fix task_2 and task_3

+++ Python_intro_6/task_3.py
@@ -1,3 +1,3 @@
-temperature_c = int(input("Введите температуру в градусах Цельсия: "))
+temperature_c = float(input("Введите температуру в градусах Цельсия: "))
 temperature_f = temperature_c * 9 / 5 + 32
 print(f"{temperature_c}°C это {temperature_f}°F")
